backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.12.3
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.9.0
FactCheckExplorer @ git+https://github.com/GONZOsint/factcheckexplorer.git@76c6ad44424e2fe14a0281ca3552117bf1e8cc69
filelock==3.17.0
Flask==3.1.0
flask-caching
flask-cors
Flask-Session==0.8.0
fonttools==4.55.8
fsspec==2025.2.0
googlesearch-python==1.3.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain-core==0.3.28
langchain-openai==0.2.14
langgraph==0.2.60
langgraph-checkpoint==2.0.10
langgraph-sdk==0.1.51
langsmith==0.2.11
MarkupSafe==3.0.2
matplotlib==3.9.4
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
networkx==3.2.1
numpy==1.26.4
openai==1.61.0
orjson==3.10.15
packaging==24.2
pillow==11.1.0
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.2.2
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wikipedia==1.4.0
youtube-transcript-api==1.2.3
